+++ front/common/utils/discussion-permissions.js
