+++ front/common/utils/variant-testing.js
