+++ front/main/app/routes/application.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import TrackClickMixin from '../mixins/track-click';
 import ArticleModel from '../models/wiki/article';
-import {activate as variantTestingActivate} from 'common/utils/variantTesting';
+import {activate as variantTestingActivate} from 'common/utils/variant-testing';
 import {normalizeToUnderscore} from 'common/utils/string';
 import Ads from 'common/modules/Ads';
 import getLinkInfo from '../utils/article-link';
